Set correct encryption mode for hardware key.
The ->encrypt method was using the wrong RSA encryption mode when
the object was representing a hardware key.  The method was
modified to use the RSA_public_entryp() function when the object
was representing a hardware key.

There were 'break' statements missing in the ->decrypt method for
the public_key and hardware_key members of the switch statement.
The break statement addition was a correctness fix for the
public_key clause and a pedantic correct for the hardware_key
member.