Add padding configuration to RSA object.
The original implementation of the object which manages use of
RSA assymetric keys assumed that Optimal Assymetric Encryption
Padding (OAEP) would be used.  This was problematic in the
presence of hardware key tokens such as the Yubikey which only
allow the use of classic PKCS1 padding.

The object state definition was modified to include a variable
which defines the type of padding which is used.  Three seperate
padding types are supported as of this update:

	NONE
	PKCS1
	PKCS1_OAEP

Object initialization sets the default padding type to be PKCS1
for compatibility reasons.

The ->set_padding method was added to change the padding type.
The only valid arguements to this method are defined in the
RSAkey.h header file as enumerated types.