Rebrand hotplug utility as device supervisor.
This commit represents a renaming of the hotplug utility to
device-supvr.

This utility is central to device initialization and setup and it
thus is installed in /sbin.  Unfortunately the kernel calls the
binary which is located at /sbin/hotplug by default on hotplug
events.  Renaming the application prevents it from being
inadvertently called by the kernel.