Relocate SmartCard object to library directory.
This update is a code relocation which moves the object which
implements support for CCID smartcards from the client directory
to the library directory.  This object will be used to support
the platform/config supervisor so this object implementation
needs to be in the library since it is referenced by multiple
systems.