Implement reader polling.
With card based CCID readers the notion of a reader and a card
are separate entities.  The smartcard API provides for concept of
polling on an insertion event to detect when a card has been
inserted into the reader.

Hardware key tokens such as the Yubikey are considered to be
readers and the notion of a card insertion event is thus not
applicable.  The ->wait_for_reader method was added to implement
polling for the presence of a new reader, which in the case of a
hardware key token such as the Yubikey, is equivalent to
detecting a card insertion event.

Support was added to the test harness to wait for a reader
presence event if the initial reader count is zero.