Update calls for RSAkey object to match API changes.
The previous changes to specification of the type of prompt
message for unlocking RSA keys broke the compilation of the
source files which are included in this update.  Add a NULL
arguement to provide the needed compatibility with the new API.
This does not affect functionality of the utilities based on
these source files since they were expecting unencrypted RSA
keys.