Add initial implementation of management supervisor daemon.
The management supervisor daemon is a process designed to be run
by the security supervisor which is responsible for managing the
local platform configuration.  The local platform configuration
is designed to be modified by the end-user and is distinct from
the device configuration which is implemented at device
provisioning and is bound to the hardware/software implementation
status.

The current implementation simply has support for detecting the
insertion of a hardware key token such as a Yubikey and using it
to decrypt the platform management token.  Success in decrypting
this token allows the end user permission to modify the
configuration of the platform.

Subsequent updates will enhance this functionality.