Add support framework for management supervision.
This commit adds the MGMTsupvr object which encapsulates the
functionality needed to implement directing a management
intervention based on the ability to decrypt a management
supervisor token with a hardware Yubikey.

The mgmt-supvr application has framework support for prompting a
user to enter a pincode and using the code to decrypt the
management token.  The encryption vector and key are printed out
on successful decryption.

The mgmt-set-key utility which is provided by this update
provides the ability to create an encrypted token which is
composed of an ASN1 encoded object which provides instructions
for how the mgmt-supvr is supposed to function.

This is closes off the work which was done to implement the EDI
transaction processor and will serve as the starting point for
subsequent development since the EDI transaction processor was
not taken into production at this version level.