Add virtual machine supervisor support.
This update implements the vm-supvr utility which is responsible
for launching a virtual machine environment in response to the
insertion of a Yubikey and the use of that Yubikey to
authenticate a platform configuration file.

Also included is a shell script which is actually responsible for
initiating the virtual machine session and managin the insertion
of the graphics adapter and USB devices needed by the virtual
machine.