Implement utility to copy monitor EDID blocks.
The iGVT (Intel Graphics Virtualization) infrastructure requires
support for 'hot-plugging' the monitor when a graphics
virtualization instance is created.  This requires copying the
128 byte Extended Display Identification Data (EDID) data block
from the system monitor to the connector instance for the
virtualized graphics adapter.

This very simplistic utility is designed to replace the use of dd
which Intel supplied scripts use.