Updates to virtual machine launch environment.
This commit is somewhat of a grab bag and is meant to synchronize
the source repository with the state of the code which was used
to implement the first FireNode demonstration unit for NRECA.

The configuration file used to gate the configuration definition
was set to the following value:

/etc/conf/config.enc

Which will need to be made configurable in subsequent updates.

The vm-supvr utility was modified so all references to the
subordinate binaries which it uses were changed to use the
standard filesystem heirarchy.

The vgt-vm script which manages the virtual machine launch
environment was modified so that it could be run by the dash
shell.  The script now contains support for creating and
releasing the LVM snapshot image which will be used as the block
device image for the guest.