Cleanup OpenSSL SSL/crypto library definitions.
Previously a library definition to bring in the objects from the
OpenSSL library was using a setup which provided access to both
the ssl and crypto libraries.  This was excessively complex and
introduced library dependencies on the ssl library which were
unneeded.

The changes implemented in this commit began the process of
isolating out a library setup for the ssl library vs a library
definition for the crypto library.  The Makefiles were cleaned up
so that utilities which only need the cryptographic routines are
just including the crypto library.