Implement library support for TPM2 hardware.
The changes implemented in this commit provide a HurdLib object
which is architected similarly to its TPM1.2 counterpart which
implements basic TPM2 functionality.  At the point of this commit
the support is preliminary.

The object definitions were headed to the NAAAIM.h at a slightly
earlier time but due to issues with respect to management of
history were left there to avoid disruptions to the change
history.