Initial implements of SGX diagnostic programs.
This update provides the implementation of two utilites named
sgx-test and sgx-metadata which provide basic functionality
testing for SGX.  The sgx-test utility creates an enclave and
adds pages to it and then destroys the enclave.  At this point
the initialization on the enclave is not performed secondary to a
lack of enclave signing.

The sgx-metadata program loads the SGX metadata section from an
enclave and outputs a summary of that information.