Encapsulate SGX metadata handling into an object.
This update introduces the SGXmetadata object which encapsulates
reading of the SGX metadata from the ELF .note.  This object also
implements the ability to dump out all of the metadata including
the section layout and patch information.

The sgx-metadata utility was modified to use this object file as
the basis for its funtionality.