Save patch and layout information in metadata.
This commit implements the collection of the patch and layout
information from the SGX enclave metadata and the saving of that
data in the state information of the object.

The patch data is an array which contains source, destination and
size data for a series of modifications which are made to the
enclave file after it is memory mapped and before it is loaded
into the enclave.

The layout data specifies the other data pages other then TEXT
data which is loaded into the enclave.  This data is larger in
footprint size and of common values since they represent heap,
stack and thread control areas.  The layouts specify the number
of times the page is to be replicated when it is commmited to the
enclave.