Implement enclave patching.
This update adds the ->patch_enclave message to the SGX metadata
object.  This method implements traversing the patch structure
array and applying the modifications specified by the patch
structures to an address range supplied as an arguement to the
method.

The modifications needed by the method required that the
sgx-metadata utility program be modified to add an additional
include file.