Implement enclave loader object and test utility.
This update provides an initial implementation of the SGXloader
object.  This object implements functionality which is needed to
load the contents of a secured enclave in ELF shared object
format into an SGX memory envelope.

In contrast to the Intel Software Development Kit this
implementation uses functionality provided by the libelf package
to implement access to the ELF data structure in the shared
object file.