Fix destruction of SGXloader object which has not been loaded.
The ->whack method was not testing whether or not the Buffer
object which holds the program segments was initialized.  This
resulted in a null-pointer dereference when the destructor was
called on an object which had not been loaded.

The destructor was modified to check to see if the Buffer object
holding the segment structure descriptors is null before walking
the list and destroying the individual segment object.