Add export of SGX Enclave Control Structure data.
The major modification in this update is output of the SECS
information which is needed to create an enclave.  This data is
abstracted from the metadata which is loaded by this object.

This commit also provides correctness updates for the
->compute_attributes method.  These changes compose the hardware
XFRM attributes into a separate variable which is then integrated
into the XFRM attributes which are requested in the metadata.

The changes in the ->compute_attributes method suggest that there
may be an error in the Intel SDK implementation when it comes to
integrating the hardware XFRM attributes into what will be used
for the SECS attributes.