Implement the SGXenclave object.
This update introduces an object which is used for general
enclave management.  At this stage the implementation is only
a partial one which allows creation of the enclave.

Two utility programs are added.  The sgx-loader utility tests the
fucntionality of the SGXloader object while the sgx-load utility
tests the SGXenclave object.