Implement enclave destruction.
The ->whack method was updated to call the kernel ioctl which
destroys the kernel implementation of the enclave.

The start address of the enclave is captured in an object state
variable named enclave_address.  Calling into the kernel
destruction is gated by whether or not the enclave_address is
non-zero.