Added ability to specify loading of a debug enclave.
This commit implements support for specifying that the enclave
which is being loaded should be created with the debug attribute
set.

While all enclaves on Linux can currently run in debug mode only
the launch enclave needs to have the debug flag cleared in order
for the enclave to pass the attribute checks in the
SGXmetadata->compute_attributes method.