Implement framework for building enclave layouts.
The SGX metadata includes an array of structures which defines
regions of the enclave which are similar in character which need
to be constructed.  Examples of these areas are the enclave heap
array and Thread Control Structure arenas.

The SGXmetadata object is tasked to build these areas since that
object aggregates all of the necessary metadata.

Each set of pages which is built in the enclave is referred to as
a layout.  There is also support for the concept of a group which
defines a preceeding range of layouts which are to be replicated.

In the current state of development the implementation simply
prints out the relevant group and layout information for
diagnostic purposes.