Add support for creating sparse virtual addess space layouts.
This commit adds the following methods to the SGXenclave object:

->add_hole

->get_address

The first method creates a 'hole' in the enclaves virtual address
space map.  A hole is a page which does not have the
implementation of a physical page in the Enclave Page Cache.

This support was added in order to support compatability with the
Intel SDK which uses such 'holes' as an implementation for guard
pages.  These guard pages are inserted between various entities
in the enclave memory arena.  The most common examples are guard
pages inserted between the heap and stack arenas.

The ->get_address method allows the value of the current virtual
address space pointer of the enclave.  This is the address of
the next page which will be added to the enclave.