Correctness fixes to SGXloader object.
This update contains multiple correctness fixes which yield a
loader object which is now idempotent with the Intel loader.  The
notion of an idempotent loader is one which generates a load of
pages which is identical to the Intel reference implemenation.
As such it should produce a processor based enclave measurement
which will allow the enclave to be initialized.

The current code has a large amount of debug output associated
with it which either needs to be removed or conditionalized.