Implement preliminary enclave initialization support.
This commit provides an initial implemention of enclave EINIT
support.  This is the final setup step needed to create a
functionally accessible enclave.

At this point the code causes the kernel drive to return an
SGX_INVALID_MEASUREMENT error.  This is secondary to the
EINITTOKEN failing to contain the necessary setup information.