Correctness fixes to implement working ENCLS.EINIT support.
These changes add the minimal set of modifications needed to
obtain working enclave initialization support.  At this point in
the development history the loader will successfully load and
measure the Intel binary Launch Enclave (LE) which is supplied by
Intel as the libsgx_le.signed.so shared object file.

Loading of the LE technically fails with a return code of
SGX_INVALID_ATTRIBUTE (2).  This is secondary to the fact that it
is necessary to use a binary SIGSTRUCT value supplied by Intel to
initialize the LE.  The Intel SDK checks for this return code and
if loading of an LE is being conducted retries the initialization
using the binary supplied SIGSTRUCT structure.

The loader will also successfully load and measure the enclave
produced by the SampleEnclave application provided in the Software
Development Kit.  Initialization of this enclave fails with a return
code of SGX_INVALID_LICENSE which is an expected outcome since,
as a non-launch enclave, EINIT requires a token generated by an
LE.