Add support for initializing the Intel binary Launch Enclave.
This commit finalizes support for loading and sealing the Intel
binary Launch Enclave (LE).  This is a necessary pre-requisite for
loading and initializing user developed enclaves.

Initializing the LE requires that a SIGSTRUCT be passed to the
ENCLS.EINIT SGX leaf instruction.  This SIGSTRUCT was signed with
the Intel SGX key which allows the procesor to load the enclave
without an EINITTOKEN which is specific the processor which the
enclave is being run on.

The SGXsigstruct object was implemented to encapsulate the
necessary SIGSTRUCT management and handling needed to support
initialization of enclaves.  This object has the current Intel
signed SIGSTRUCT integrated into the object itself.  This
provides a simplified method of implementing this functionality
since maintaining and then loading a separate binary SIGSTRUCT
file is not needed.