Implement separate LE initialization and thread support.
In preparation implementing support to call into an enclave this
commit implements a separate method to initialize the Intel
supplied Launch Enclave.  This method uses the SIGSTRUCT supplied
in binary format in the call to the ENCLS.EINIT call.

Support is also added to save the Task Control Structures (TCS)
in the enclave.  This is needed in order to support selecting a
free TCS slot when a context of execution is dispatched into the
enclave.