Save Task Control Structure (TCS) designations as virtual addresses.
The ENCLU.EENTER instruction requires that the TCS which is to be
used for an enclave context of execution be passed into the
enclave as a virtual address which maps into the Enclave Page
Cache (EPC) page where the physical TCS structure is stored.

The previous implementation of the code which was saving thread
definitions was attempting to use a copy of the TCS which was
stored in the enclave callers non-secured virtual memory.  This
was resulting in a general protection fault (GP) secondary to the
fact the address was not within the enclave virtual address space
mapping.