Add initial enclave execution support.
This update represents the initial execution of an enclave with
the IDFusion SGX runtime environment.

The sgx-gen-token utility was introduced as a framework to execute
the Intel Launch Enclave which is released in binary form.  The
ultimate purpose of this utility is to generate an EINITTOKEN for
arbitrary enclaves.

This commit also introduces the enter_enclave.o object file.
This is a copy of an object file which was generated by compiling
the following source file in the Intel SDK:

psw/urts/linux/enter_enclave.S

This file implements the enclave entry code.  It will be
subsequently replaced by our own implementation but the object
file is included in order to provide a fully functional enclave
loader.