Surface the enclave attributes.
This commit implements a call chain which allows the computed
security attributes of an enclave to be available at the level of
the enclave object.

The enclave security attributes are one of the components which
need to be supplied to the Launch Enclave in order to generate an
EINITTOKEN for the enclave.  Under the assumption that an
external tool will carry out initialization of an enclave this
information needs to be available.