Implemented initial support for enclave initialization.
This update provides the initial framework for generating an
EINITTOKEN or launch token for an arbitrary enclave using a
loaded Launch Enclave (LE).  At this point in the development
history the sgx-gen-token utility should be complete with respect
to matching the process undertaken in the Intel SDK.

The LE requires that a certificate white list be loaded into the
enclave prior to a request being made to generate a launch token.  This
white list is provided in binary only form by Intel and is
downloaded by the system development kit as a prebuilt binary.

The gen-white-list utility is added in this update to convert
this binary file into a C structure which is included in the
sgx-gen-token utility rather then reading the certificate white
list from an external file.

At the current time both entry slots to the LE are returning an
error code of 1 from the enter_enclave call.  Work needs to be
done to determine why the call to the enclave is returning with
an unexpected error status.