Add support for enclave initialization.
The issue with the enclave entry code returning 1 was secondary
to the fact that the enclave initialization code was not being
run.  This update provides support for calling the enclave
initialization 'slot' prior to attempting to load the enclave
white list of certificates.

The execution slots in an enclave begin numbering with 0 for the
application defined enclave routines.  The -1 slot is common to
all enclaves and is used to initialize the enclave for execution.
The initialization code is supplied by the Trusted RunTime System
(TRTS) and includes functionality such as carrying out any
necessary reloations in the enclave.

At this point in development the sgx-gen-token utility will now
successfully create an EINITTOKEN for arbitrary enclaves.