Implement SGXenclave->boot_slot method.
This update implements a method for calling an enclave execution
slot.  It integrates a great deal of code from the sgx-gen-token
utility where the enclave execution code had been prototyped.

At the curent time thread management is decidedly simplistic but
it is now encapsulated in the enclave object as well.  The
->boot_slot method decrements the current thread count manually
in order to compensate for the reference count increase which is
implemented in the ->get_thread method.