Integrate enclave initialization thread into SGXenclave object.
This update adds support for calling the initial enclave
initialization process on a loaded enclave.  A separate private
method was added which implements the call to the -1 function
slot and a call to this method was added to the ->init_enclave
and ->init_launch_enclave methods.

For simplicity a null pointer is passed to the ->boot_slot code
for the thread manager.  Since no OCALL API definition table is
provided the assumption is that it is safe to not specify a
thread handler object for an outcall.