Add support for saving EINITTOKEN.
The Intel AESM runtime environment saves the launch token for a
given enclave in a file named 'enclave.token' in the executing
user's home directory.  Subsequent invocations of the enclave use
this cached token.

This update modifies the sgx-gen-token utility to save a
generated token in a file named 'enclave.token' in the current
directory of the application when it is executed.

The EINITTOKEN is currently 512 bytes long and is aligned on a
512 byte boundary.  For some reason the Intel application uses a
1024 byte buffer to save the EINITTOKEN.  This buffer is saved
which leaves the effective token file size as 1K rather then the
actual 512 bytes of the EINITTOKEN structure.  The sgx-gen-token
utility emulates this behavior.

With this update we can demonstrate the ability to execute the
SampleEnclave application without the AESM service running by
first generating an enclave.token with the sgx-gen-token utility.