Implement generic sgx-gen-token improvements.
This update adds getopt processing of command-line arguements to
increase program flexibility.

This update also gates the output of debugging information with
the -d debug switch.

Default values are implemented for the location of the launch
enclave, output to stdout and the location of the character
device node used to access SGX kernel services.