Implement replacement enclave trampoline code.
The major focus of this update is the implementation of the
boot-sgx.s assembler file which implements the code which
mediates the call from the SGXenclave->boot_slot method into the
enclave itself.

The boot-sgx.s file also provides the interface which will field
an OCALL from the enclave and vector that into the correct
userspace routines which implement the OCALL.

With this code in place the loader system is no longer dependent
on the enter_enclave.o object file which had been garnered from
the Intel Software Development Kit.  That file will be deleted
after support is added for properly vectoring the OCALL requests.