Implement platform specific enclave initialization.
This update modifies the SGXenclave->init_enclave method to
accept an arguement which is a pointer to an EINITTOKEN which has
been generated for this platform.  This processor specific token
is required for any enclave which does not have the launch
attribute set on it.

The sgx-load utility was modified to use getopt based
command-line arguements for specifying the parameters to be used
for the load.

The utility was also updated to load the launch token from a user
specified file.