Stack correctness fixes for enclave bootcode.
Work on implementing support for OCALL's from an enclave
uncovered correctness problems with stack handling in the enclave
boot code.

The substraction which is designed to align the stack pointer on
a 16-byte boundary was attempting to substract -8 bytes from the
stack rather then decrementing the stack pointer by 8 bytes to
take the stack pointer to the next 16-byte boundary.  Fix this
and change the substraction to a quad-word substraction.

The code which was restoring the values which were pushed onto
the stack by the entry code was using offsets which were much too
large and were resulting in corruption of the contents of the
stack.  The mov instructions were modified to have the correct
offsets from the RBP pointer.