Implement enclave OCALL support.
This commit represents a major update which implements support
for an enclave to request processing of a function in the
untrusted portion of the application by an enclave.

The trampoline code in boot-sgx.s was modified to implement a
call to the sgx-ocall function which bridges the OCALL return
back into the enclave object which executed the ECALL into the
enclave.  There is no technical reason for the sgx-ocall function
to exist except for simplicity with respect to being able to call
a method within the object definition structure.

The ->boot_ocall method was added to the SGXenclave object which
is invoked by the previously mentioned sgx_ocall function.  The
boot_ocall method uses the slot reference provided to it be the
OCALL to pick en ECALL resolution function from the ECALL API
set.

The test-ecall application was added to demonstrate support for
this functionality.  The binary enclave which is needed by this
application is not currently included but will be added in a
subsequent update.