Implement proper OCALL enclave return.
The enclave trampoline code was not properly re-entering the
enclave after an OCALL was processed.  This update issues an
ENCLU.EENTER command after successful OCALL processing in order
to return to executing in the context of the enclave.

A total of three arguements which are passed to the boot_sgx
function need to be preserved to support OCALL processing.  This
update moves saving of those values to immediately after the
registers which must be preserved over the function call.

Since a total of three arguements are preserved the subtraction
of eight bytes from the stack pointer can be removed since the
stack pointer on entry to the enclave is now on a 16-byte
boundary.