Implement user input looping in test-ecall utility.
The test-ecall utility was modified to accept user input in a
loop and to echo that input through the companion enclave.

The updated version of the enclave is also included in this
commit.  The updated enclave issues a second printk in order to
test the ability of the enclave to execute multiple OCALL's.