Update to new TPM2 TSS library API.
A base upgrade to the Wok development environment was implemented
which upgraded major core elements of the library which included
the version of the MUSL C library, OpenSSL and the IBMTSS TPM2
API library.

The upgrade to version 851 of the IBMTSS library included a major
API change.  This update implements the necessary build changes
to allow the tpm2-cmd to build with the new library interface.

One of the primary differences was to rename the library prefix
to TSS from TSE.  There was also a change to the NVram public API
structure which involved the removal of a union definition.

The linker flags used in the util sub-directory were changed to
se the following directive:

-Wl,-rpath /usr/local/musl/lib

Which causes a path directive to be included in each binary so
the MUSL based shared libraries can be located.

This required that the load-image linkage command be updated to
include the following directive:

-Wl,-rpath-link /usr/local/musl/lib

Which specifies the library location the dynamic linker should
search in order to resolve shared library dependencies.