Implement memory based enclave loading and initialization.
This series of changes completes support for loading and
initializing an enclave image which is held in memory rather then
in the form of a shared object file.

The primary modifications were made in the SGXloader object which
adds the following two methods:

->load_memory

->load_secs_memory

Which are the memory based equivalents of the ->load and
->load_secs methods.

The two new methods take as arguements a pointer to a memory
buffer which contains the image of an enclave and a size value
which designates the size of of the image in memory.

The SGXenclave object has the following additional method:

->load_memory

Which takes a memory based enclave image and size as its
arguements for the enclave to be managed.

While not included in this update these new methods have been
tested with a modified version of the sgx-gen-token utility which
has the Intel Launch Enclave loaded from an in memory based
image.