Add missing object variable definitions.
The SGXenclave object was missing the initialization of its
debug object variable which would cause an object to randomly
implement debugging mode.  An explicit initialization to false
was implemented.

The more problematic missing initialization was the fd variable
in the SGXloader object.  This would be randomly initialized to
zero which caused the file descriptor for stdin to be randomly
closed which required multiple hours of debugging to find.