Add 'Fire And Forget' demonstration binaries.
This update implements the ability to build the following two
binaries:

idf-sgx-fandf

idf-sgx-fandf-unibin

These are examples of how the IDfusion SGX runtime support
environment is capable of created SGX unified binaries.  Unified
binaries are an SGX enabled application which has all of the
functionality, including the enclave itself, encapsulated into a
single binary.

These binaries only require an SGX kernel driver on the platform
which they are deployed to.  There is no need for the Intel PSW
software to be installed or accessible.

The source code is designed to have the launch enclave and
optionally the execution enclave be available as include files.
These include files are not currently committed into the source
repository but can be generated with the gen-white-list utility.
The intention is to modify this utility to be a generic tool for
generating not only the white list but launch and application
enclaves as well.