Implement initial trajectory mapping support.
The ISO-identity directory had been previously initialized to
hold the patches for various versions of the kernel code which
implements the iso-identity behavioral modeling code.  This
directory will also contain the trajectory mapping code which
will be used to generate behavior contour maps to be processed by
the kernel modeling code.

The trajectory mapping code was started with the Actor object
which encapsulates the processing of an actor identity.  A
testing utility called test-actor was implemented to exercise the
basic functionality of this object.