Implement a Subject object.
This update provides an initial implementation of a subject
object which encapsulates and measures the identity elements of a
subject node.  It is the object which is complementary to the
Actor object which was previously committed.

A small utility program called test-actor is also part of this
commit which is used to exercise the basic functionality of the
Subject object.