Implement contour generation utility.
This commit provides an initial implementation of a utility which
traces an execution trajectory map and generates iso-identity
contours and optionally debug information about the map.

The 'generate-contours' utility takes two arguements.  The -i
switch is used to specify the name of the trajectory file while
the -v switch enables full debug information.

The contours which are generated without the -v switch are
suitable for writing into the following pseudo-file:

/sys/kernel/security/ima/iso-identity/map

Which initializes the iso-identity behavioral state of the
system.