Initial iso-identity modifications to the boot supervisor.
Supporting the iso-identity modeling system requires some basic
changes to the boot supervisor.

The securityfs filesystem needs to be mounted for manipulation of
the system model.  The do_mounts() function was modified to
include this pseudo-filesystem as the normal set of filesystems
to be mounted and unmounted by the system initializer.

If a request to seal the system is noted by the presence of a
root password file the boot supervisor issues a sync() call to
make sure the sealing files are flushed to disk.  A sleep command
is allowed for this operation to complete and then an explicit
system halt is requested.

The do_reboot() function was modified to accept a arguement which
specifies whether the system is to shutdown or reboot.