Add support for loading system behavior contours.
In order to implement a stable measurement state the previous
version of the boot supervisor would read all of the files which
were included in a manifest located in the /boot/MANIFEST file of
the loaded filesystem.

The supervisor was modified to read a set of behavior contours
which are read from the /boot/contours file.  These contours are
written to the /sys/kernel/security/ima/iso-identity/map file to
initialize the behavioral model of the system.