TPM daemon correctness fixes.
The ordering of shutdown of the boot supervisor was inconsistent.
The filesystems were unmounted before shutdown of the daemon was
requested.  Change the ordering so the TPM daemon shutdown occurs
first following by the unmount of all filesystems.

The boot supervisor was naively sending a SIGTERM signal to the
TPM management daemon.  Since the boot supervisor was running as
PID 1 and did not wait for the daemon to shutdown this left the
first execution of the daemon as a zombie.  After the termination
signal is sent the supervisor now attempts to read the child for
five seconds before giving up and signaling an error.