Change update of TPM persistent storage to write only.
The opening of the /mnt/var/lib/tpm/system.data file in
read/write mode was causing an open writers violation to be
flagged.  The mode for the update was changed to write only to
defeat the violation.  This eliminates the need to flag the file
location as a pseudonym.

This raises the question of whether or not this is a possible
measurement gap since a process with root privileges could
technically inject a file into the system without being measured.
Any attempt to use the file, ie. opening it in readable mode with
invoke measurement so technically the file could not be accessed
without a behavioral assessment of it being caught.