Implement initialization of system behavior.
The boot supervisor now reads the /boot/pseudonyms and
/boot/contours files in order to configure platform behavior
status.

Implement initialization of system behavior.

The boot supervisor now reads the /boot/pseudonyms and
/boot/contours files in order to configure platform behavior
status.

No action is taken if the files are not present.  If the files
are present their content is read and the entres are committed
to the following configuration pseudo-files:

/boot/pseudonyms: /sys/kernel/security/ima/iso-identity/pseudonyms

/boot/contours: /sys/kernel/security/ima/iso-identity/map

Pseudonyms are used to define files whose contents are to be
considered invariant and thus not generate either TOMTOU or open
writers violations.

Contours define dimensional points in the platform behavior
space.  They represent step points in the approved execution
trajectory path.