Add system measurement tool.
The compute-measurement utility added in this update implements
the computation of a gross system measurement based on the
following input:

- Boot aggregate measurement.

- Host identity.

- Trajectory contours.

Without the verbose arguement the utility simply outputs the SHA256
measurement of the described system in ASCII hexadecimal format.

Placng the utility in verbose mode with the -v command-line
arguement outputs all of the computation steps involved in the
calculation of the system measurement.