Implement test program for set_behavior system call.
A utitility named set-behavior was created which provides a test
harness for testing the functionality of the sys_behavior system
call which has been added to the iso-identity modeling system in
the Linux kernel.

This utility currently takes two arguements which are specified
with the -c and -p command line options.  The -c arguement
specifies that a behavior contour point is to be added to the
current system behavior model.  The arguement to the -c option is
the contour point in ASCII hexadecimal format.

The -p arguement specifies that a file pseudonym is to be
registered.  The arguement to the -p option is the full path
description of the pseudonym.