Convert boot supervisor to use sys_behavior system call.
This update converts configuration of the behavior pseudonyms and
contour points from the sysfs filesystem interface to the system
call interface.  This change also eliminates the need for the
sleep calls which had been placed after the configurations in
order to address the race condition associated with the kernel
asynchronously feeding the configuration options into the
modeling engine.