Add namespace functionality to set-behavior utility.
The iso-identity modeling engine supports the notion of
behavioral namespaces which allow a process to request an
instance of a behavioral domain for itself and all subordinate
processes.  A behavioral namespace is requested via the
specification of the CLONE_BEHAVIOR bit which is analagous to the
bits in the CLONE_* flags which request instances of other
operating system resources.

This update adds the -r and -n switches to the set-behavior
utility to request either a new namespace (-n) or to run a
command in a namespace (-r).  The -r switch requires an arguement
to specify the strings which is to be passed to the system() call
to be run in an independent namespace.